Aca se trata de implementar lo del JWT, marca un error en authController.js, se modifico el tema de los status hhtp, y unas partes de las validaciones.

+++ backend-compromiso/src/api/proceso/controllers/procesoController.js
@@ -20,10 +20,10 @@ const getProcesos = async (req, res) => {
 const getProcesoById = async (req, res) => {
   try {
     const proceso = await Proceso.findByPk(req.params.id);
-    if (proceso) {
-      res.json(proceso);
-    } else {
+    if (!proceso) {
       res.status(404).json({ message: 'Proceso no encontrado' });
+    } else {
+      res.json(proceso);
     }
   } catch (error) {
     logger.error(error.message, { stack: error.stack });
@@ -31,21 +31,46 @@ const getProcesoById = async (req, res) => {
   }
 };
 
+
+
 const createProceso = async (req, res) => {
   try {
-    const proceso = await Proceso.create(req.body);
+    const { nombre_proceso } = req.body;
+
+    // Validar que el campo requerido esté presente
+    if (!nombre_proceso) {
+      return res.status(400).json({ error: 'Datos requeridos faltantes o inválidos' });
+    }
+
+    // Crear el proceso si la validación es correcta
+    const proceso = await Proceso.create({ nombre_proceso });
     res.status(201).json(proceso);
   } catch (error) {
     logger.error(error.message, { stack: error.stack });
-    res.status(500).json({ error: 'Error interno del servidor' });
+    
+    // Manejo de errores con un mensaje genérico
+    if (error.name === 'SequelizeValidationError' || error.name === 'SequelizeUniqueConstraintError') {
+      return res.status(400).json({ error: 'Error en el procesamiento de datos' });
+    }
+    return res.status(500).json({ error: 'Error interno del servidor' });
   }
 };
 
+
+
+
 const updateProceso = async (req, res) => {
   try {
+    const proceso = await Proceso.findByPk(req.params.id);
+
+    if (!proceso) {
+      return res.status(404).json({ message: 'Proceso no encontrado' });
+    }
+
     const [updated] = await Proceso.update(req.body, {
       where: { Id_Proceso: req.params.id },
     });
+
     if (updated) {
       const updatedProceso = await Proceso.findByPk(req.params.id);
       res.json(updatedProceso);
@@ -58,11 +83,21 @@ const updateProceso = async (req, res) => {
   }
 };
 
+
+
+
 const deleteProceso = async (req, res) => {
   try {
+    const proceso = await Proceso.findByPk(req.params.id);
+
+    if (!proceso) {
+      return res.status(404).json({ message: 'Proceso no encontrado' });
+    }
+
     const deleted = await Proceso.destroy({
       where: { Id_Proceso: req.params.id },
     });
+
     if (deleted) {
       res.status(204).end();
     } else {
